Don't remove the control file if it wasn't a temp copy



git-svn-id: svn://test.kernel.org/autotest/trunk@1064 592f7852-d20e-0410-864c-8624ca9c26a4

+++ server/autotest.py
@@ -195,7 +195,8 @@ class Autotest(installable_object.InstallableObject):
 		# Copy control_file to remote_control_file on the host
 		tmppath = utils.get(control_file)
 		host.send_file(tmppath, atrun.remote_control_file)
-		os.remove(tmppath)
+		if os.path.abspath(tmppath) != os.path.abspath(control_file):
+			os.remove(tmppath)
 
 		try:
 			atrun.execute_control(timeout=timeout)
